requirements.txt
pandas==2.1.4
pyarrow==17.0.0
psycopg2-binary==2.9.9
apache-airflow-providers-http==4.13.1
apache-airflow-providers-amazon==8.28.0
apache-airflow-providers-postgres==5.13.1